# openAPI_Spec/admin端接口.openapi.yaml
openapi: 3.1.0
info:
  title: admin端接口
  description: ''
  version: 1.0.0
...
paths:
...
  /admin/setmeal:
    put:
      summary: 修改套餐
      deprecated: false
      description: ''
      tags:
        - 套餐相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: integer
                  format: int64
                  description: 分类id
                description:
                  type: string
                  description: 套餐描述
                id:
                  type: integer
                  format: int64
                  description: 套餐id
                image:
                  type: string
                  description: 套餐图片路径
                name:
                  type: string
                  description: 套餐名称
                price:
                  type: number
                  description: 套餐价格
                setmealDishes:
                  type: array
                  items:
                    type: object
                    properties:
                      copies:
                        type: integer
                        format: int32
                        description: 菜品份数
                      dishId:
                        type: integer
                        format: int64
                        description: 菜品id
                      id:
                        type: integer
                        format: int64
                        description: 关系表主键值
                      name:
                        type: string
                        description: 菜品名称
                      price:
                        type: number
                        description: 菜品价格
                      setmealId:
                        type: integer
                        format: int64
                        description: 套餐id
                    title: SetmealDish
                    $$ref: '#/definitions/SetmealDish'
                    required:
                      - copies
                      - dishId
                      - name
                      - price
                  description: 套餐和菜品关联关系
                status:
                  type: integer
                  format: int32
                  description: 套餐起售状态
              title: SetmealDTO
              $$ref: '#/definitions/SetmealDTO'
              required:
                - categoryId
                - id
                - image
                - name
                - price
                - setmealDishes
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties: {}
                  msg:
                    type: string
                title: R
                $$ref: '#/definitions/R'
                required:
                  - code
                  - data
          headers: {}
      security: []
    delete:
      summary: 批量删除套餐
      deprecated: false
      description: ''
      tags:
        - 套餐相关接口
      parameters:
        - name: ids
          in: query
          description: ids
          required: true
          example: ''
          schema:
            type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties: {}
                  msg:
                    type: string
                title: R
                $$ref: '#/definitions/R'
          headers: {}
      security: []
    post:
      summary: 新增套餐
      deprecated: false
      description: ''
      tags:
        - 套餐相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: integer
                  format: int64
                  description: 分类id
                description:
                  type: string
                  description: 套餐描述
                id:
                  type: integer
                  format: int64
                  description: 套餐id
                image:
                  type: string
                  description: 套餐图片
                name:
                  type: string
                  description: 套餐名称
                price:
                  type: number
                  description: 套餐价格
                setmealDishes:
                  type: array
                  items:
                    type: object
                    properties:
                      copies:
                        type: integer
                        format: int32
                        description: 份数
                      dishId:
                        type: integer
                        format: int64
                        description: 菜品id
                      id:
                        type: integer
                        format: int64
                        description: 套餐和菜品关系id
                      name:
                        type: string
                        description: 菜品名称
                      price:
                        type: number
                        description: 菜品价格
                      setmealId:
                        type: integer
                        format: int64
                        description: 套餐id
                    title: SetmealDish
                    $$ref: '#/definitions/SetmealDish'
                    required:
                      - copies
                      - dishId
                      - name
                      - price
                      - setmealId
                  description: 套餐包含的菜品
                status:
                  type: integer
                  format: int32
                  description: 套餐状态：1位起售 0为停售
              title: SetmealDTO
              $$ref: '#/definitions/SetmealDTO'
              required:
                - categoryId
                - image
                - name
                - price
                - setmealDishes
                - status
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties: {}
                  msg:
                    type: string
                title: R
                $$ref: '#/definitions/R'
                required:
                  - code
          headers: {}
      security: []
  /admin/setmeal/page:
    get:
      summary: 分页查询
      deprecated: false
      description: ''
      tags:
        - 套餐相关接口
      parameters:
        - name: categoryId
          in: query
          description: 分类id
          required: false
          example: ''
          schema:
            type: string
        - name: name
          in: query
          description: 套餐名称
          required: false
          example: ''
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          required: true
          example: ''
          schema:
            type: string
        - name: pageSize
          in: query
          description: 每页记录数
          required: true
          example: ''
          schema:
            type: string
        - name: status
          in: query
          description: 套餐起售状态
          required: false
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                  msg:
                    type: 'null'
                  data:
                    type: object
                    properties:
                      total:
                        type: number
                      records:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: number
                            categoryId:
                              type: number
                            name:
                              type: string
                            price:
                              type: number
                            status:
                              type: number
                            description:
                              type: string
                            image:
                              type: string
                            updateTime:
                              type: string
                            categoryName:
                              type: string
                          required:
                            - id
                            - categoryId
                            - name
                            - price
                            - status
                            - description
                            - image
                            - updateTime
                            - categoryName
                required:
                  - code
          headers: {}
      security: []
  /admin/setmeal/status/{status}:
    post:
      summary: 套餐起售、停售
      deprecated: false
      description: ''
      tags:
        - 套餐相关接口
      parameters:
        - name: status
          in: path
          description: 套餐状态，1表示起售，0表示停售
          required: true
          example: '1'
          schema:
            type: string
        - name: id
          in: query
          description: 套餐id
          required: true
          example: '101'
          schema:
            type: string
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties: {}
                  msg:
                    type: string
                title: R
                $$ref: '#/definitions/R'
                required:
                  - code
          headers: {}
      security: []
  /admin/setmeal/{id}:
    get:
      summary: 根据id查询套餐
      deprecated: false
      description: ''
      tags:
        - 套餐相关接口
      parameters:
        - name: id
          in: path
          description: 套餐id
          required: true
          example: '101'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      categoryId:
                        type: integer
                        format: int64
                      categoryName:
                        type: string
                      description:
                        type: string
                      id:
                        type: integer
                        format: int64
                      image:
                        type: string
                      name:
                        type: string
                      price:
                        type: number
                      setmealDishes:
                        type: array
                        items:
                          type: object
                          properties:
                            copies:
                              type: integer
                              format: int32
                            dishId:
                              type: integer
                              format: int64
                            id:
                              type: integer
                              format: int64
                            name:
                              type: string
                            price:
                              type: number
                            setmealId:
                              type: integer
                              format: int64
                          title: SetmealDish
                          $$ref: '#/definitions/SetmealDish'
                          required:
                            - copies
                            - dishId
                            - id
                            - name
                            - price
                            - setmealId
                      status:
                        type: integer
                        format: int32
                      updateTime:
                        type: string
                        format: date-time
                    title: SetmealVO
                    $$ref: '#/definitions/SetmealVO'
                    required:
                      - categoryId
                      - categoryName
                      - description
                      - id
                      - image
                      - name
                      - price
                      - setmealDishes
                      - status
                      - updateTime
                  msg:
                    type: string
                title: R«SetmealVO»
                $$ref: '#/definitions/R«SetmealVO»'
                required:
                  - code
                  - data
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers: []
security: []